refactor: package foundry bluegreen tool
