+++ tools/foundry_bluegreen_pkg/errors.py
@@ -1,0 +1,7 @@
+"""Foundry blue/green workflow exceptions."""
+
+from __future__ import annotations
+
+
+class FoundryBlueGreenError(RuntimeError):
+    """Raised when a workflow precondition fails."""
